Pull queue builder out of daemon
