+++ src/utils/mod.rs
@@ -6,6 +6,7 @@ pub use self::daemon::start_daemon;
 pub use self::github_updater::github_updater;
 pub use self::html::extract_head_and_body;
 pub use self::queue::{get_crate_priority, remove_crate_priority, set_crate_priority};
+pub use self::queue_builder::queue_builder;
 pub use self::release_activity_updater::update_release_activity;
 pub(crate) use self::rustc_version::parse_rustc_version;
 
@@ -19,6 +20,7 @@ mod github_updater;
 mod html;
 mod pubsubhubbub;
 mod queue;
+mod queue_builder;
 mod release_activity_updater;
 mod rustc_version;
 pub(crate) mod sized_buffer;
